runpod/image/requirements.txt
runpod
torch
# diffusers from source — required for ZImagePipeline (merged in latest release)
git+https://github.com/huggingface/diffusers
transformers
accelerate
safetensors
Pillow
omegaconf
huggingface_hub
requests
xformers
